Use Redux Toolkit
- 기존 액션-액션함수-리덕스의 중복을 제거하기 위해 createSlice() 사용
- 그 외 깨지는 테스트 케이스 수정
