+++ src/reducer.test.js
@@ -1,6 +1,4 @@
-import reducer from './reducer';
-
-import {
+import reducer, {
   setRegions,
   setCategories,
   setRestaurants,
@@ -13,7 +11,8 @@ import {
   changeReviewField,
   clearReviewFields,
   setReviews,
-} from './actions';
+} from './reducer';
+
 
 describe('reducer', () => {
   context('when previous state is undefined', () => {
