+++ src/store.js
@@ -1,9 +1,10 @@
-import { createStore, applyMiddleware } from 'redux';
-
-import thunk from 'redux-thunk';
+import { configureStore } from '@reduxjs/toolkit';
 
 import reducer from './reducer';
 
-const store = createStore(reducer, applyMiddleware(thunk));
+const store = configureStore({
+  reducer,
+  devTools: process.env.NODE_ENV !== 'production',
+});
 
 export default store;
